+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 Breaking
-  * AZ-717: Bump azurerm provider to v3
+  * AZ-717: Bump AzureRM provider version to `v3.0+`
 
 # v5.1.0 - 2022-04-08